requirements.txt
mcp
requests
pandas
google-adk